Create analyze text
These functions analyze a text or body of word for use in my other poetry functions. They make it possible to experiment with the different tones that come out of writing poems using only words from The Dark Side of the Moon or Alice's Adventures in Wonderland, for example. Running the mgroups2 function on the text is all you need to do. 
It would be a lot better with more accurate POS tagging though- preferably tagging that is more accurate and more specific, differentiating between transitive verbs and intransitive verbs, for example.