metered function
The meter2 function returns a metered line of poetry. The first input is the type of meter, and the second input is how many 'feet' you want. The "maxsyl" input is set to 3 by default and should be set to the number of syllables in the word in your group of words with the highest number of syllables. Then input which part of speech (NOT in string form) you want each word in the line to be. For example, for a line of iambic pentameter with four words- subject, transitive verb, object, adverb- your inputs  would be " 'iambic', 5, maxsyl, subjects, transitives, objects, adverbs ". if you only have one and two syllable words available, for example, you should set maxsyl as 2. In this version, if you call for a word to be a part of speech that does not have any words that fit the meter, the function will replace it with ANY word that fits the meter.