poem generators
The functions build on all of the functions in the other files to generate poems. Once you have analyzed your text (and defined all of the other functions), you can use it to generate some poems. 
So far, the poems are good for showing the tone of a text used as the word source, but they make no sense. Hopefully, I can develop better syntax via better and more specific POS tagging.